+++ src/GenericTensorNetworks.jl
@@ -16,11 +16,11 @@ using Primes
 using DocStringExtensions
 using Base.Cartesian
 using ProblemReductions
-import ProblemReductions: ConstraintSatisfactionProblem, AbstractSatisfiabilityProblem, UnitWeight
+import ProblemReductions: ConstraintSatisfactionProblem, AbstractSatisfiabilityProblem, UnitWeight, hard_constraints, is_satisfied, local_solution_spec, solution_size, energy_mode, LargerSizeIsBetter, SmallerSizeIsBetter
 import ProblemReductions: @bv_str, StaticElementVector, StaticBitVector, onehotv, _nints, hamming_distance
 import ProblemReductions: is_set_covering, is_vertex_coloring, is_set_packing, is_dominating_set, is_matching, is_maximal_independent_set, cut_size, is_independent_set
 import ProblemReductions: num_paint_shop_color_switch, spin_glass_from_matrix, CNF, CNFClause, BoolVar, satisfiable, @bools, ∨, ¬, ∧
-import ProblemReductions: flavors, set_weights, weights, num_flavors, variables
+import ProblemReductions: flavors, set_weights, weights, is_weighted, num_flavors, variables, energy
 import AbstractTrees: children, printnode, print_tree
 import StatsBase
 
@@ -40,18 +40,23 @@ export CountingTropicalF64, CountingTropicalF32, TropicalF64, TropicalF32, Exten
 export generate_samples, OnehotVec
 
 # Graphs
+export HyperGraph, SimpleGraph, UnitDiskGraph
 export random_regular_graph, diagonal_coupled_graph
-export square_lattice_graph, unit_disk_graph, random_diagonal_coupled_graph, random_square_lattice_graph
+export square_lattice_graph, random_diagonal_coupled_graph, random_square_lattice_graph
 export line_graph
 
+
+# Problems
+export AbstractProblem, ConstraintSatisfactionProblem
+export hard_constraints, is_satisfied, local_solution_spec, solution_size, energy_mode, LargerSizeIsBetter, SmallerSizeIsBetter
 # Tensor Networks (Graph problems)
 export GenericTensorNetwork, optimize_code, UnitWeight
-export flavors, variables, num_flavors, weights, fixedvertices, set_weights, energy_terms
+export flavors, variables, num_flavors, weights, fixedvertices, set_weights, is_weighted
 export IndependentSet, mis_compactify!, is_independent_set
 export MaximalIS, is_maximal_independent_set
 export cut_size, MaxCut
-export spinglass_energy, spin_glass_from_matrix, SpinGlass
-export PaintShop, paintshop_from_pairs, num_paint_shop_color_switch, paint_shop_coloring_from_config, paint_shop_from_pairs
+export energy, spin_glass_from_matrix, SpinGlass
+export PaintShop, paintshop_from_pairs, num_paint_shop_color_switch
 export Coloring, is_vertex_coloring
 export Satisfiability, CNF, CNFClause, BoolVar, satisfiable, @bools, ∨, ¬, ∧
 export DominatingSet, is_dominating_set